# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GSNpeeps API
  version: 0.7.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Kontrak HTTP GSNpeeps untuk Employee Database, dashboard HR, absensi,
    approval, notifikasi, dan administrasi akses.

    API menggunakan field JSON `snake_case`, Bearer JWT berumur 8 jam, dan
    cross-check sesi aktif di Redis. Tidak ada refresh-token endpoint.
  contact:
    name: GSNpeeps IT Liaison
servers:
  - url: http://{host}:{port}
    description: Development
    variables:
      host:
        default: localhost
      port:
        default: "8080"
  - url: https://{host}
    description: Production
    variables:
      host:
        default: api.janjikupadamu.id
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/ketidakhadiran:
    post:
      operationId: createLeaveRequest
      summary: Ajukan ketidakhadiran
      description: |
        Karyawan, Atasan, dan HR mengajukan cuti/izin/perjalanan dinas milik
        sendiri. Dokumen pendukung wajib. Sistem menentukan approver awal menurut
        role/struktur dan membuat notifikasi idempotent untuk approver tersebut.
      tags: [Leave Requests]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateLeaveRequest"
      responses:
        "201":
          description: Pengajuan berhasil dibuat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequestCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/FileTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedFileType"
        "422":
          description: Saldo cuti atau kuota jenis izin tidak memenuhi aturan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                insufficientLeaveBalance:
                  value:
                    success: false
                    error:
                      code: INSUFFICIENT_LEAVE_BALANCE
                      message: Saldo cuti tidak mencukupi
                quotaExceeded:
                  value:
                    success: false
                    error:
                      code: QUOTA_EXCEEDED
                      message: Durasi melebihi kuota jenis izin
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listLeaveRequestsForApproval
      summary: Ambil antrean approval ketidakhadiran
      description: |
        Atasan hanya melihat bawahan langsung dan tahap aktifnya. HR hanya melihat
        pengajuan pada kewenangan HR. Top Management hanya melihat pengajuan HR.
      tags: [Leave Requests]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApprovalQueueStatus"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Antrean pengajuan sesuai scope approver.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequestListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/master/jenis-izin:
    get:
      operationId: listLeaveTypes
      summary: Ambil master jenis izin
      description: >-
        Terbuka untuk seluruh role terautentikasi karena pemohon membutuhkan
        jenis_izin_id saat mengajukan ketidakhadiran (resolusi G-012). Hanya HR yang
        menerima master nonaktif; role lain selalu dibatasi ke jenis izin aktif dan
        parameter aktif diabaikan. Administrasi master tetap HR-only.
      tags: [Leave Types]
      security:
        - BearerAuth: []
      parameters:
        - name: aktif
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Daftar jenis izin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveTypeListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createLeaveType
      summary: Tambah jenis izin
      description: Hanya HR. Perubahan dicatat pada Audit Log.
      tags: [Leave Types]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveTypeRequest"
      responses:
        "201":
          description: Jenis izin berhasil dibuat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/master/jenis-izin/{id}:
    put:
      operationId: updateLeaveType
      summary: Perbarui jenis izin
      description: Hanya HR. Dapat mengubah kuota atau menonaktifkan master.
      tags: [Leave Types]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeaveTypeIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveTypeRequest"
      responses:
        "200":
          description: Jenis izin berhasil diperbarui.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token berlaku 8 jam. Backend juga memverifikasi session aktif di
        Redis pada setiap request terproteksi.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
...
    LeaveTypeIdPath:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  responses:
    BadRequest:
      description: Request atau parameter tidak valid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            invalidRequest:
              value:
                success: false
                error: { code: INVALID_REQUEST, message: Request tidak valid }
    Unauthorized:
      description: Kredensial tidak ada, tidak valid, atau kedaluwarsa.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unauthorized:
              value:
                success: false
                error: { code: UNAUTHORIZED, message: Autentikasi diperlukan }
    Forbidden:
      description: User tidak memiliki kewenangan atau akses row-level.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            forbidden:
              value:
                success: false
                error: { code: FORBIDDEN, message: Akses ditolak }
    NotFound:
      description: Resource tidak ditemukan.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            notFound:
              value:
                success: false
                error: { code: NOT_FOUND, message: Data tidak ditemukan }
    Conflict:
      description: Konflik state atau data duplikat.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            conflict:
              value:
                success: false
                error: { code: CONFLICT, message: Data sudah ada atau sudah diproses }
...
    FileTooLarge:
      description: File melebihi batas 5 MB.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            fileTooLarge:
              value:
                success: false
                error: { code: FILE_TOO_LARGE, message: Ukuran file maksimal 5 MB }
    UnsupportedFileType:
      description: Tipe file tidak didukung.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unsupported:
              value:
                success: false
                error: { code: UNSUPPORTED_FILE_TYPE, message: Tipe file tidak didukung }
    TooManyRequests:
      description: Rate limit terlampaui atau akun terkunci sementara.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            rateLimited:
              value:
                success: false
                error: { code: RATE_LIMITED, message: Terlalu banyak request }
    InternalError:
      description: Kesalahan internal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            internal:
              value:
                success: false
                error: { code: INTERNAL_ERROR, message: Terjadi kesalahan internal }
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 0193f406-8d2f-7d0c-bf34-61ff192e46de
...
    RequestStatus:
      type: string
      description: Enum machine-readable yang digunakan sama pada database, query, dan response.
      enum: [menunggu_atasan, menunggu_hr, menunggu_top_management, disetujui, ditolak, dibatalkan]
    ApprovalQueueStatus:
      type: string
      enum: [menunggu_atasan, menunggu_hr, menunggu_top_management]
...
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: Validasi gagal }
        fields:
          $ref: "#/components/schemas/ValidationFields"
    ValidationFields:
      type: object
      additionalProperties: { type: string }
      example: { email: Format email tidak valid }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error: { $ref: "#/components/schemas/ErrorObject" }
    PaginationMeta:
      type: object
      required: [page, limit, total_data, total_page]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total_data: { type: integer, minimum: 0 }
        total_page: { type: integer, minimum: 0 }
    IdData:
      type: object
      required: [id]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
    IdSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/IdData" }
    RequestStateData:
      type: object
      required: [id, status]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        status: { $ref: "#/components/schemas/RequestStatus" }
    LeaveRequestCreateResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/RequestStateData" }
...
    LeaveType:
      type: object
      required: [id, kode, nama, kuota_tahunan, is_active]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        kode: { type: string, maxLength: 50 }
        nama: { type: string, maxLength: 150 }
        kuota_tahunan: { type: integer, minimum: 0 }
        memerlukan_dokumen: { type: boolean }
        is_active: { type: boolean }
    CreateLeaveTypeRequest:
      type: object
      additionalProperties: false
      required: [kode, nama, kuota_tahunan, memerlukan_dokumen]
      properties:
        kode: { type: string, maxLength: 50 }
        nama: { type: string, maxLength: 150 }
        kuota_tahunan: { type: integer, minimum: 0 }
        memerlukan_dokumen: { type: boolean }
    UpdateLeaveTypeRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nama: { type: string, maxLength: 150 }
        kuota_tahunan: { type: integer, minimum: 0 }
        memerlukan_dokumen: { type: boolean }
        is_active: { type: boolean }
    LeaveTypeListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/LeaveType" } }
    CreateLeaveRequest:
      type: object
      required: [jenis_izin_id, tanggal_mulai, tanggal_selesai, alasan]
      properties:
        jenis_izin_id: { $ref: "#/components/schemas/UUID" }
        tanggal_mulai: { type: string, format: date }
        tanggal_selesai: { type: string, format: date }
        alasan: { type: string, minLength: 10, maxLength: 2000 }
        dokumen_pendukung:
          type: string
          format: binary
          description: Wajib bila master jenis izin mensyaratkannya; maksimal 5 MB.
        lokasi_tujuan:
          type: string
          description: Wajib secara kondisional untuk Perjalanan Dinas.
        keterangan_lokasi:
          type: string
          description: Wajib secara kondisional untuk Perjalanan Dinas.
...
    LeaveRequestSummary:
      type: object
      required:
        [id, employee_id, nama_karyawan, jenis_izin, tanggal_mulai,
         tanggal_selesai, jumlah_hari, status, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        employee_id: { $ref: "#/components/schemas/UUID" }
        nama_karyawan: { type: string }
        jenis_izin: { type: string }
        tanggal_mulai: { type: string, format: date }
        tanggal_selesai: { type: string, format: date }
        jumlah_hari: { type: integer, minimum: 1 }
        status: { $ref: "#/components/schemas/RequestStatus" }
        created_at: { type: string, format: date-time }
...
    LeaveRequestListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          required: [meta]
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/LeaveRequestSummary" } }
            meta: { $ref: "#/components/schemas/PaginationMeta" }
...